c2cgeoform/scaffolds/c2cgeoform/requirements.txt
#-e git+https://github.com/camptocamp/c2cgeoform#egg=c2cgeoform
ColanderAlchemy==0.3.3
plaster_pastedeploy
pyramid >= 1.9a
pyramid_debugtoolbar
pyramid_retry
pyramid_tm
pyramid_jinja2
SQLAlchemy
transaction
zope.sqlalchemy
waitress
-e ../c2cgeoform
-e .
